makes responsive

+++ src/components/Expenses/ExpenseDate.js
@@ -1,7 +1,7 @@
 import './ExpenseDate.css';
 
 function ExpenseDate(props) {
-  const month = props.date.toLocaleString('en-US', { month: 'long' });
+  const month = props.date.toLocaleString('en-US', { month: 'short' });
   const day = props.date.toLocaleString('en-US', { day: '2-digit' });
   const year = props.date.getFullYear();
